Renaming of GVON and Graph
